refactor: Schedule 도메인 의존성 제거
